+++ packages/intellij-client/src/main/kotlin/com/blazejkustra/reactcompilermarker/actions/GenerateReportAction.kt
@@ -49,6 +49,7 @@ class GenerateReportAction : AnAction() {
             "headExtra" to headExtra,
             "excludeDirs" to settings.excludedDirectoriesList,
             "includeExtensions" to settings.supportedExtensionsList,
+            "respectGitignore" to settings.respectGitignore,
             "emojis" to mapOf(
                 "success" to settings.successEmoji,
                 "error" to settings.errorEmoji